refactor(web): move changelog/roadmap content to repo root, drop mono all-caps labels
CHANGELOG.md and ROADMAP.md now live at the repo root (the conventional,
discoverable location) instead of apps/web/src/content/; the pages import them
from there via Vite ?raw. Eyebrow, date, tag, and section labels switch from
mono uppercase to capitalized sans.
