+++ apps/web/package.json
@@ -8,7 +8,7 @@
     "dev": "vite",
     "build": "tsc -b && vite build",
     "preview": "vite preview",
-    "test": "tsx --test src/*.test.ts src/*/*.test.ts",
+    "test": "tsx --test src/content/*.test.ts",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
